zcash: Reduce size of the signed PCZT before returning it
The unsigned PCZT provided by the wallet contains various optional
fields that we use to both check the PCZT effects and create the
signatures. However, we can assume that the wallet can recover these
fields if it needs them, and clear them from the PCZT we return. This
should make the animated QR code displayed by the Keystone device
smaller and easier to scan.